+++ src/UI/Component/Item/Factory.php
@@ -37,46 +37,6 @@ interface Factory
      * ---
      * description:
      *   purpose: >
-     *     Contributions are quotes that can be credited to one explicit contributor and time and which have a focus on
-     *     that contributor, since the origin of the quote declaring its value.
-     *     Contribution Items are used to present that quote with a clear visible association to its creator. This can
-     *     be used to present comments, reactions ,references or ratings and more.
-     *   composition: >
-     *     Contributions contain a textual quote, a contributor name and a datetime presentation or a representative
-     *     info if the datetime or contributor of the creation is unknown.
-     *     Further the Contribution can have a close button and a lead icon.
-     *   effect: >
-     *     An interaction with the close button may remove the Contribution permanently.
-     * rivals: >
-     *     Standard Items have their value inside the content itself, while their creators are insignificant for the
-     *     benefiting group of observers. Contributions draw their worth form the contributor. They present content
-     *     which is benefiting for the observer because it was created by this contributor.
-     * rules:
-     *   interaction:
-     *     1: >
-     *        Clicking on the Close Button MUST remove the Contribution Item permanently.
-     *   accessibility:
-     *     1: >
-     *       All interactions offered by a Contribution Item MUST be accessible by only using the keyboard.
-     *     2: >
-     *       The main quote of the contribution MUST NOT be part of any interaction.
-     * ---
-     *
-     * @param string      $quote
-     * @param string      $contributor
-     * @param \ilDateTime $createDatetime
-     * @return \ILIAS\UI\Component\Item\Contribution
-     */
-    public function contribution(
-        string $quote,
-        ?string $contributor = null,
-        ?DateTimeImmutable $createDatetime = null
-    ) : Contribution;
-
-    /**
-     * ---
-     * description:
-     *   purpose: >
      *       An Item Group groups items of a certain type.
      *   composition: >
      *       An Item Group consists of a header with an optional action Dropdown and
