+++ src/UI/examples/Contribution/Quote/base.php
@@ -1,18 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace ILIAS\UI\examples\Item\Contribution;
+namespace ILIAS\UI\examples\Contribution\Quote;
 
-function with_close()
+function base()
 {
     global $DIC;
 
     return $DIC->ui()->renderer()->render(
-        $DIC->ui()->factory()->item()->contribution(
+        $DIC->ui()->factory()->contribution()->quote(
             'a little test contribution',
             'Contributor',
             new \DateTimeImmutable()
-        )->withClose(
-            $DIC->ui()->factory()->button()->close()
         )
     );
 }
