+++ src/UI/examples/Contribution/Quote/with_lead_icon.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace ILIAS\UI\examples\Item\Contribution;
+namespace ILIAS\UI\examples\Contribution\Quote;
 
 use ILIAS\UI\Component\Symbol\Icon\Standard;
 
@@ -9,7 +9,7 @@ function with_lead_icon()
     global $DIC;
 
     return $DIC->ui()->renderer()->render(
-        $DIC->ui()->factory()->item()->contribution(
+        $DIC->ui()->factory()->contribution()->quote(
             'a little test contribution',
             'Contributor',
             new \DateTimeImmutable()
