+++ src/UI/examples/Contribution/Quote/with_long_content.php
@@ -1,13 +1,13 @@
 <?php declare(strict_types=1);
 
-namespace ILIAS\UI\examples\Item\Contribution;
+namespace ILIAS\UI\examples\Contribution\Quote;
 
 function with_long_content()
 {
     global $DIC;
 
     return $DIC->ui()->renderer()->render(
-        $DIC->ui()->factory()->item()->contribution(
+        $DIC->ui()->factory()->contribution()->quote(
             'This is a test contribution of the root user in the current time. This is a little bit longer than
             the usual to test its visual presentation when the content exceed a minor amount of chars to see if this
             is still presented properly. This may affect its presentation inside a mobile or resticted view hterefore
