+++ src/UI/examples/Contribution/Quote/with_unkown_contributor.php
@@ -1,13 +1,13 @@
 <?php declare(strict_types=1);
 
-namespace ILIAS\UI\examples\Item\Contribution;
+namespace ILIAS\UI\examples\Contribution\Quote;
 
 function with_unkown_contributor()
 {
     global $DIC;
 
     return $DIC->ui()->renderer()->render(
-        $DIC->ui()->factory()->item()->contribution(
+        $DIC->ui()->factory()->contribution()->quote(
             'a little test contribution',
         )
     );
