--- src/UI/examples/Item/Contribution/base.php
@@ -1,16 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ILIAS\UI\examples\Item\Contribution;
-
-function base()
-{
-    global $DIC;
-
-    return $DIC->ui()->renderer()->render(
-        $DIC->ui()->factory()->item()->contribution(
-            'a little test contribution',
-            'Contributor',
-            new \DateTimeImmutable()
-        )
-    );
-}
